requirements.txt
django
djangorestframework
drf-yasg
geopy
django-filter==24.3
celery

